+++ CHANGELOG.md
@@ -23,6 +23,9 @@
 * Added throttling to the `ProgressBar` update methods to impact the actual process' performance as little as possible.
 * Added a new class `Spinner` to the `console` module, which is used to display a spinner animation in the console during an ongoing process.
 
+**BREAKING CHANGES:**
+* Made the value input into the params `bar_format` and `limited_bar_format` of `ProgressBar` be a list/tuple of strings instead of a single string, so the user can define multiple formats for different console widths.
+* Added a new param <code>sep: str = " "</code> to the `ProgressBar` class, which is used to join multiple bar-format strings.
 
 <span id="v1-9-0" />