+++ CHANGELOG.md
@@ -25,9 +25,10 @@
 
 **BREAKING CHANGES:**
 * Made the value input into the params `bar_format` and `limited_bar_format` of `ProgressBar` be a list/tuple of strings instead of a single string, so the user can define multiple formats for different console widths.
-* Added a new param <code>sep: str = " "</code> to the `ProgressBar` class, which is used to join multiple bar-format strings.
+* Added a new param <code>sep: *str* = " "</code> to the `ProgressBar` class, which is used to join multiple bar-format strings.
 * Renamed the class property `Console.wh` to `Console.size`, since it describes the property better.
 * Renamed the class property `Console.usr` to `Console.user`, since it describes the property better.
+* Added missing type checking to methods in the `path` module.
 
 <span id="v1-9-0" />